.pdc $alt -> $cruise
$alt is wrong because:
$alt
The temporary or the current altitude (if no temporary is specified) of the selected aircraft.
 $cruise
The final cruise altitude of the selected aircraft.
source:
http://www.dutchflightnetwork.nl/images/PDF_Documenten/euroscopemanual.pdf
TLDR: $alt can conflict with the temp alt if one is set to an arriving aircraft that stay connected and file another FP, where as $cruise wont